V3 Prototype: Firebase + Material-UI
In this commit we introduce a working prototype of CoursePlanner v3, based on material-ui for UI changes and Firestore DB for persistent storage. We also use Firestore for authentication, currently using only the Google Sign-In provider.

With this change, users can now login with their Google account and save schedules. Firebase + Firestore takes care of authentication and authorization so there is no backend server.

Prototype is not currently hosted on the courseplanner domain.

TODO: minor UI changes, update saved schedule, delete saved schedule.