Notify user on successful save/deletion.
This change triggers a Snackbar notification message indicating a successful save or deletion.

Note: this could be improved to handle multiple back-to-back deletions better. Only one snackbar can be shown at a time, and this implementation doesn't handle automatically closing an old Snackbar to display a new Snackbar. Deletions are pretty visible by seeing the schedule disappear from the list of saved schedules, so I'm not going to worry about this for now.