Allow deleting individual courses.
This change adds a small delete icon next to each course row in the course table. If all courses in a saved schedule are deleted, we clear the current session as if the user cleared the entire schedule.

One small but strange consequence: some calendar colors will change with deletions. This is because colors aren't assigned to specific courses, they're assigned to indices in the list of courses which get modified when courses are deleted. This could probably be improved in the future.